// src/app/openapi-gen/model/userCreationDto.ts
/**
 * GratisBrockiBackend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: @springdoc.version@
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserCreationDto { 
    username?: string;
    email?: string;
    password?: string;
    phoneNumber?: string;
}